More menu work with a new file structure. Change game slot with 'select' is now working. Normalized game flow with system_return()
