+++ include/service.h
@@ -4,6 +4,7 @@
 #include <stdint.h>
 
 #define SERVICE_CURSOR          ((volatile int8_t *)  0x10FCDA)
+#define SERVICE_CURSOR_MAX      ((volatile uint8_t *) 0x10FCDB)
 #define SERVICE_CALENDAR_CURSOR ((volatile uint8_t *) 0x10FCDA)
 #define SERVICE_LED_MARQUEE     ((volatile uint8_t *) 0x10FCDA)
 #define SERVICE_LED_COIN        ((volatile uint8_t *) 0x10FCDE)
@@ -20,237 +21,34 @@
 #define NEW_BIOS_MINUTE         ((volatile uint8_t *) 0x10FCE7)
 #define NEW_BIOS_SECOND         ((volatile uint8_t *) 0x10FCE8)
 
-// Hardware Test
-void show_crosshatch_test();
-void show_color_test();
-void show_io_test();
-void update_io_test();
-void show_sound_test();
-void update_sound_test();
-void show_memory_card_test();
-void update_memory_card_test();
-void show_backup_clear();
-void update_backup_clear();
-void show_setup_calendar();
-void update_setup_calendar();
+// BIOS menu
+#define MENU_BIOS_MAIN          8
+#define MENU_BIOS_HARDWARE      9
+#define MENU_BIOS_HARD_DIPS     10
+#define MENU_BIOS_SOFT_DIPS     11
+#define MENU_BIOS_BOOK_KEEPING  12
+#define MENU_BIOS_CODE_NUMBER   13
+#define MENU_BIOS_CALENDAR      14
+#define MENU_BIOS_EXIT          15
 
-// BIOS Menu
+// Hardware menu
+#define MAX_NUM_MENUS       7
+#define MENU_CROSSHATCH     0
+#define MENU_COLOR          1
+#define MENU_IO             2
+#define MENU_SOUND_TEST     3
+#define MENU_MEMORY_CARD    4
+#define MENU_CLEAR_BACKUP   5
+#define MENU_SETUP_CALENDAR 6
+
 void show_bios_menu();
-void show_bios_menu_hard_dips();
-void update_bios_menu_hard_dips();
-void show_bios_menu_soft_dips();
-void update_bios_menu_soft_dips();
-void show_bios_menu_game_soft_dips();
+void show_bios_main_menu();
+void update_bios_main_menu();
+void show_bios_hardware_test();
 
 // Local service functions
 volatile uint16_t * _add_large_char(volatile uint16_t *address, uint8_t selected, uint16_t position, uint16_t value);
 void _move_cursor(uint8_t max_menu_items);
-
-static const uint16_t red_green_blue [] = {
-    0x0003, 0x7069, 
-    0x0108, 0x2020, 0x5245, 0x4420, 0x2020, 0x2047, 0x5245, 0x454e, 0x2020, 0x2020, 0x2042, 0x4c55, 0x4520, 0x2020, 0x2057, 0x4849, 0x5445, 0xffff, 
-    0x0003, 0x706d, 
-    0x1108, 0x2020, 0x2020, 0x2020, 0xffff, 0x0003, 0x706f, 0x1108, 0x2020, 0x2020, 0x2020, 0xffff, 
-    0x0003, 0x7071, 
-    0x1108, 0x2020, 0x2020, 0x2020, 0xffff, 
-    0x0003, 0x718d, 
-    0x2108, 0x2020, 0x2020, 0x2020, 0xffff, 
-    0x0003, 0x718f, 
-    0x2108, 0x2020, 0x2020, 0x2020, 0xffff, 
-    0x0003, 0x7191, 
-    0x2108, 0x2020, 0x2020, 0x2020, 0xffff, 
-    0x0003, 0x72ad, 
-    0x3108, 0x2020, 0x2020, 0x2020, 0xffff, 
-    0x0003, 0x72af, 
-    0x3108, 0x2020, 0x2020, 0x2020, 0xffff, 
-    0x0003, 0x72b1, 
-    0x3108, 0x2020, 0x2020, 0x2020, 0xffff, 
-    0x0003, 0x73cd, 
-    0x4108, 0x2020, 0x2020, 0x2020, 0xffff, 
-    0x0003, 0x73cf, 
-    0x4108, 0x2020, 0x2020, 0x2020, 0xffff, 
-    0x0003, 0x73d1, 
-    0x4108, 0x2020, 0x2020, 0x2020, 0xffff, 
-    0x0000
-};
-
-static const uint16_t io [] = {
-    0x0001, 0x00ff, 
-    0x2002, 
-    0x0003, 0x71e5, 
-    0x0007, 0x492f, 0x4f20, 0x4348, 0x4543, 0x4bff, 
-    0x0003, 0x7187, 
-    0x0007, 0x5031, 0x2050, 0x32ff, 
-    0x0003, 0x70c9, 
-    0x0007, 0x5550, 0xffff, 
-    0x0005, 0x0001, 
-    0x0007, 0x444f, 0x574e, 0xffff, 
-    0x0005, 0x0001, 
-    0x0007, 0x4c45, 0x4654, 0xffff, 
-    0x0005, 0x0001, 
-    0x0007, 0x5249, 0x4748, 0x54ff, 
-    0x0005, 0x0001, 
-    0x0007, 0x5055, 0x5348, 0x2041, 0xffff, 
-    0x0005, 0x0001, 
-    0x0007, 0x5055, 0x5348, 0x2042, 0xffff, 
-    0x0005, 0x0001, 
-    0x0007, 0x5055, 0x5348, 0x2043, 0xffff, 
-    0x0005, 0x0001, 
-    0x0007, 0x5055, 0x5348, 0x2044, 0xffff,
-    0x0005, 0x0001, 
-    0x0007, 0x5354, 0x4152, 0x54ff, 
-    0x0005, 0x0001, 
-    0x0007, 0x5345, 0x4c45, 0x4354, 0xffff, 
-    0x0003, 0x7327, 
-    0x0007, 0x3132, 0x3334, 0x3536, 0x3738, 0xffff, 
-    0x0003, 0x7288, 
-    0x0007, 0x4449, 0x50ff, 
-    0x0003, 0x728b,
-    0x0007, 0x5445, 0x5354, 0xffff, 
-    0x0005, 0x0001, 
-    0x0007, 0x434f, 0x494e, 0x31ff, 
-    0x0005, 0x0001, 
-    0x0007, 0x434f, 0x494e, 0x32ff, 
-    0x0005, 0x0001, 
-    0x0007, 0x5345, 0x5256, 0x4943, 0x45ff, 
-    0x0003, 0x7293, 
-    0x0007, 0x454c, 0x204c, 0x4544, 0x3120, 0x4c45, 0x4432, 0xffff,
-    0x0000
-};
-
-static const uint16_t io_values [] = {
-    // EL
-    0x0101, 0x0002,
-    0x2002,
-    0x0003, 0x72b4,
-    0x0004, 0x0010, 0xfc00,
-
-    // LED 1
-    0x0003, 0x7314,
-    0x0006,
-
-    // LED 2
-    0x0003, 0x73b4,
-    0x0006,
-
-    // P1 / P2
-    0x0101, 0x000a,
-    0x0102,
-    0x0003, 0x71a9,
-    0x0006,
-    0x0003, 0x7209,
-    0x0006, 
-
-    // Dip switches
-    0x0101, 0x0008,
-    0x2002,
-    0x0003, 0x7328,
-    0x0006,
-
-    // Test, coin and service
-    0x0101, 0x0004,
-    0x0102,
-    0x0003, 0x73ab,
-    0x0006,
-
-    0x0000
-};
-
-static const uint16_t sound [] = {
-    0x0003, 0x71c6, 
-    0x0108, 0x534f, 0x554e, 0x4420, 0x5445, 0x5354, 0xffff,
-    0x000a, 0x0010, 0xfc00,
-    0x0000, 0x0000
-};
-
-static const char * const sound_menu[] = {
-    "NO SOUND",
-    "RIGHT",
-    "LEFT",
-    "CENTER"
-};
-
-static const uint16_t memory_card [] = {
-    0x0003, 0x7186,
-    0x0108, 0x4d45, 0x4d4f, 0x5259, 0x2043, 0x4152, 0x4420, 0x5445, 0x5354, 0xffff,
-    0x0003, 0x714c,
-    0x1108, 0x1149, 0x4e53, 0x4552, 0x5420, 0x4d45, 0x4d4f, 0x5259, 0x2043, 0x4152, 0x4410, 0xffff, 
-    0x0000
-};
-
-static const uint16_t memory_card_ok [] = {
-    0x0003, 0x70ec, 
-    0x0108, 0x4d45, 0x4d4f, 0x5259, 0x2043, 0x4152, 0x4420, 0x5445, 0x5354, 0x204f, 0x4bff,
-    0x0000
-};
-
-static const uint16_t service_memory_card_error [] = {
-    0x0003, 0x70ec,
-    0x0108, 0x4d45, 0x4d4f, 0x5259, 0x2043, 0x4152, 0x4420, 0x5445, 0x5354, 0x2045, 0x5252, 0x4f52, 0xffff, 
-    0x0000
-};
-
-static const uint16_t backup_clear [] = {
-    // 'backup clear'
-    0x0003, 0x71c6,
-    0x0108, 0x4241, 0x434b, 0x5550, 0x2043, 0x4c45, 0x4152, 0xffff, 
-    // 'ok'
-    0x0003, 0x710c,
-    0x1108, 0x4f4b, 0xffff,
-    // ' = press a,b,c buttons'
-    0x0003, 0x714c,
-    0x0108, 0x203d, 0x2050, 0x5245, 0x5353, 0x2041, 0x2c42, 0x2c43, 0x2042, 0x5554, 0x544f, 0x4e53, 0xffff, 
-    // '     at the same time'
-    0x0003, 0x710e,
-    0x0108, 0x2020, 0x2020, 0x2041, 0x5420, 0x5448, 0x4520, 0x5341, 0x4d45, 0x2054, 0x494d, 0x45ff, 
-    0x0000
-};
-
-static const uint16_t service_setup_calendar [] = {
-    // 'setting up the calendar'
-    0x0003, 0x7106, 
-    0x0108, 0x5345, 0x5454, 0x494e, 0x4720, 0x5550, 0x2054, 0x4845, 0x2043, 0x414c, 0x454e, 0x4441, 0x52ff, 
-
-    // 'current time'
-    0x0003, 0x7169,
-    0x0108, 0x4355, 0x5252, 0x454e, 0x5420, 0x5449, 0x4d45, 0xffff, 
-    
-    // Add /
-    0x0003, 0x71ac,
-    0x0108, 0x2f20, 0x202f, 0xffff,
-
-    // Add :
-    0x0003, 0x72cc,
-    0x0108, 0x3a20, 0x203a, 0xffff, 
-
-    // 'new time'
-    0x0003, 0x7170, 
-    0x0108, 0x4e45, 0x5720, 0x5449, 0x4d45, 0xffff,
-
-    // Add / and : 
-    0x0003, 0x71b3,
-    0x0108, 0x2f20, 0x202f, 0x2020, 0x2020, 0x203a, 0x2020, 0x3aff,
-
-    // 'joystick = select'
-    0x0003, 0x7176,
-    0x0108, 0x4a4f, 0x5953, 0x5449, 0x434b, 0x203d, 0x2053, 0x454c, 0x4543, 0x54ff,
-
-    // 'd button = set'
-    0x0003, 0x7179,
-    0x0108, 0x4420, 0x4255, 0x5454, 0x4f4e, 0x203d, 0x2053, 0x4156, 0x45ff, 
-
-    0x0000
-};
-
-static const char *weekdays [] = {
-    "SUN", "MON", "TUE", "WED", "THU", "FRI", "SAT"
-};
-
-static const uint32_t service_setup_current_values [][2] = {
-    {(uint32_t)0x720a, (uint32_t)BIOS_YEAR},
-    {(uint32_t)0x714a, (uint32_t)BIOS_MONTH},
-    {(uint32_t)0x71aa, (uint32_t)BIOS_DAY},
-};
 
 static const uint16_t bios_menu [] = {
     0x0003, 0x7145, 
@@ -269,65 +67,6 @@ static const uint16_t bios_menu [] = {
     0x0108, 0x4558, 0x4954, 0xffff, 
     0x0003, 0x725a,
     0x0108, 0x2f20, 0x202f, 0x2020, 0xffff, 
-    0x0000
-};
-
-static const uint16_t bios_menu_hard_dips [] = {
-    // 'setting up the hard dips'
-    0x0003, 0x7105,
-    0x0108, 0x5345, 0x5454, 0x494e, 0x4720, 0x5550, 0x2054, 0x4845, 0x2048, 0x4152, 0x4420, 0x4449, 0x5053, 0xffff,
-    // '123456789'
-    0x0003, 0x7249,
-    0x0108, 0x3132, 0x3334, 0x3536, 0x3738, 0xffff,
-    // '0=off 1=on'
-    0x0003, 0x7259, 
-    0x0108, 0x303d, 0x4f46, 0x4620, 0x313d, 0x4f4e, 0xffff,
-    // 'setting mode'
-    0x0003, 0x708c,
-    0x0108, 0x5345, 0x5454, 0x494e, 0x4720, 0x4d4f, 0x4445, 0xffff,
-    // 'controller'
-    0x0005, 0x0001,
-    0x0108, 0x434f, 0x4e54, 0x524f, 0x4c4c, 0x4552, 0xffff,
-    // 'chute'
-    0x0005, 0x0001,
-    0x0108, 0x4348, 0x5554, 0x45ff,
-    // 'communication'
-    0x0005, 0x0001,
-    0x0108, 0x434f, 0x4d4d, 0x554e, 0x4943, 0x4154, 0x494f, 0x4eff,
-    // 'free play'
-    0x0005, 0x0001,
-    0x0108, 0x4652, 0x4545, 0x2050, 0x4c41, 0x59ff,
-    // 'freeze mode'
-    0x0005, 0x0001,
-    0x0108, 0x4652, 0x4545, 0x5a45, 0x204d, 0x4f44, 0x45ff,
-    // '........'
-    0x0003, 0x724c,
-    0x0108, 0x2e2e, 0x2e2e, 0x2e2e, 0x2e2e, 0xffff,
-    0x0005, 0x0001,
-    0x0108, 0x2e2e, 0x2e2e, 0x2e2e, 0x2e2e, 0xffff,
-    0x0005, 0x0001,
-    0x0108, 0x2e2e, 0x2e2e, 0x2e2e, 0x2e2e, 0xffff,
-    0x0005, 0x0001,
-    0x0108, 0x2e2e, 0x2e2e, 0x2e2e, 0x2e2e, 0xffff,
-    0x0005, 0x0001,
-    0x0108, 0x2e2e, 0x2e2e, 0x2e2e, 0x2e2e, 0xffff,
-    0x0005, 0x0001,
-    0x0108, 0x2e2e, 0x2e2e, 0x2e2e, 0x2e2e, 0xffff,
-    0x0000
-};
-
-// Assumes active low, ON = 0, OFF = 1
-static const char *bios_menu_on_off [] = {
-    "ON ", "OFF"
-};
-
-static const char *bios_menu_communication_link [] = {
-    "[No 1]", "[No 2]", "[No 3]", "[No 4]",
-};
-
-static const uint16_t bios_menu_soft_dips [] = {
-    0x0003, 0x7145,
-    0x0108, 0x5345, 0x5454, 0x494e, 0x4720, 0x5550, 0x2054, 0x4845, 0x2043, 0x4142, 0x494e, 0x4554, 0xffff,
     0x0000
 };
