+++ include/service_dips.h
@@ -1,0 +1,71 @@
+#ifndef _SERVICE_DIPS_H
+#define _SERVICE_DIPS_H
+
+#include <stdint.h>
+
+void show_bios_menu_hard_dips();
+void update_bios_menu_hard_dips();
+void show_bios_menu_soft_dips();
+void update_bios_menu_soft_dips();
+void show_bios_menu_game_soft_dips();
+
+static const uint16_t bios_menu_hard_dips [] = {
+    // 'setting up the hard dips'
+    0x0003, 0x7105,
+    0x0108, 0x5345, 0x5454, 0x494e, 0x4720, 0x5550, 0x2054, 0x4845, 0x2048, 0x4152, 0x4420, 0x4449, 0x5053, 0xffff,
+    // '123456789'
+    0x0003, 0x7249,
+    0x0108, 0x3132, 0x3334, 0x3536, 0x3738, 0xffff,
+    // '0=off 1=on'
+    0x0003, 0x7259, 
+    0x0108, 0x303d, 0x4f46, 0x4620, 0x313d, 0x4f4e, 0xffff,
+    // 'setting mode'
+    0x0003, 0x708c,
+    0x0108, 0x5345, 0x5454, 0x494e, 0x4720, 0x4d4f, 0x4445, 0xffff,
+    // 'controller'
+    0x0005, 0x0001,
+    0x0108, 0x434f, 0x4e54, 0x524f, 0x4c4c, 0x4552, 0xffff,
+    // 'chute'
+    0x0005, 0x0001,
+    0x0108, 0x4348, 0x5554, 0x45ff,
+    // 'communication'
+    0x0005, 0x0001,
+    0x0108, 0x434f, 0x4d4d, 0x554e, 0x4943, 0x4154, 0x494f, 0x4eff,
+    // 'free play'
+    0x0005, 0x0001,
+    0x0108, 0x4652, 0x4545, 0x2050, 0x4c41, 0x59ff,
+    // 'freeze mode'
+    0x0005, 0x0001,
+    0x0108, 0x4652, 0x4545, 0x5a45, 0x204d, 0x4f44, 0x45ff,
+    // '........'
+    0x0003, 0x724c,
+    0x0108, 0x2e2e, 0x2e2e, 0x2e2e, 0x2e2e, 0xffff,
+    0x0005, 0x0001,
+    0x0108, 0x2e2e, 0x2e2e, 0x2e2e, 0x2e2e, 0xffff,
+    0x0005, 0x0001,
+    0x0108, 0x2e2e, 0x2e2e, 0x2e2e, 0x2e2e, 0xffff,
+    0x0005, 0x0001,
+    0x0108, 0x2e2e, 0x2e2e, 0x2e2e, 0x2e2e, 0xffff,
+    0x0005, 0x0001,
+    0x0108, 0x2e2e, 0x2e2e, 0x2e2e, 0x2e2e, 0xffff,
+    0x0005, 0x0001,
+    0x0108, 0x2e2e, 0x2e2e, 0x2e2e, 0x2e2e, 0xffff,
+    0x0000
+};
+
+// Assumes active low, ON = 0, OFF = 1
+static const char *bios_menu_on_off [] = {
+    "ON ", "OFF"
+};
+
+static const char *bios_menu_communication_link [] = {
+    "[No 1]", "[No 2]", "[No 3]", "[No 4]",
+};
+
+static const uint16_t bios_menu_soft_dips [] = {
+    0x0003, 0x7145,
+    0x0108, 0x5345, 0x5454, 0x494e, 0x4720, 0x5550, 0x2054, 0x4845, 0x2043, 0x4142, 0x494e, 0x4554, 0xffff,
+    0x0000
+};
+
+#endif // _SERVICE_DIPS_H
