+++ include/service_hardware.h
@@ -1,0 +1,213 @@
+#ifndef _SERVICE_HARDWARE_H
+#define _SERVICE_HARDWARE_H
+
+void show_crosshatch_test();
+void show_color_test();
+void show_io_test();
+void update_io_test();
+void show_sound_test();
+void update_sound_test();
+void show_memory_card_test();
+void update_memory_card_test();
+void show_backup_clear();
+void update_backup_clear();
+void show_setup_calendar();
+void update_setup_calendar();
+
+static const uint16_t red_green_blue [] = {
+    0x0003, 0x7069, 
+    0x0108, 0x2020, 0x5245, 0x4420, 0x2020, 0x2047, 0x5245, 0x454e, 0x2020, 0x2020, 0x2042, 0x4c55, 0x4520, 0x2020, 0x2057, 0x4849, 0x5445, 0xffff, 
+    0x0003, 0x706d, 
+    0x1108, 0x2020, 0x2020, 0x2020, 0xffff, 0x0003, 0x706f, 0x1108, 0x2020, 0x2020, 0x2020, 0xffff, 
+    0x0003, 0x7071, 
+    0x1108, 0x2020, 0x2020, 0x2020, 0xffff, 
+    0x0003, 0x718d, 
+    0x2108, 0x2020, 0x2020, 0x2020, 0xffff, 
+    0x0003, 0x718f, 
+    0x2108, 0x2020, 0x2020, 0x2020, 0xffff, 
+    0x0003, 0x7191, 
+    0x2108, 0x2020, 0x2020, 0x2020, 0xffff, 
+    0x0003, 0x72ad, 
+    0x3108, 0x2020, 0x2020, 0x2020, 0xffff, 
+    0x0003, 0x72af, 
+    0x3108, 0x2020, 0x2020, 0x2020, 0xffff, 
+    0x0003, 0x72b1, 
+    0x3108, 0x2020, 0x2020, 0x2020, 0xffff, 
+    0x0003, 0x73cd, 
+    0x4108, 0x2020, 0x2020, 0x2020, 0xffff, 
+    0x0003, 0x73cf, 
+    0x4108, 0x2020, 0x2020, 0x2020, 0xffff, 
+    0x0003, 0x73d1, 
+    0x4108, 0x2020, 0x2020, 0x2020, 0xffff, 
+    0x0000
+};
+
+static const uint16_t io [] = {
+    0x0001, 0x00ff, 
+    0x2002, 
+    0x0003, 0x71e5, 
+    0x0007, 0x492f, 0x4f20, 0x4348, 0x4543, 0x4bff, 
+    0x0003, 0x7187, 
+    0x0007, 0x5031, 0x2050, 0x32ff, 
+    0x0003, 0x70c9, 
+    0x0007, 0x5550, 0xffff, 
+    0x0005, 0x0001, 
+    0x0007, 0x444f, 0x574e, 0xffff, 
+    0x0005, 0x0001, 
+    0x0007, 0x4c45, 0x4654, 0xffff, 
+    0x0005, 0x0001, 
+    0x0007, 0x5249, 0x4748, 0x54ff, 
+    0x0005, 0x0001, 
+    0x0007, 0x5055, 0x5348, 0x2041, 0xffff, 
+    0x0005, 0x0001, 
+    0x0007, 0x5055, 0x5348, 0x2042, 0xffff, 
+    0x0005, 0x0001, 
+    0x0007, 0x5055, 0x5348, 0x2043, 0xffff, 
+    0x0005, 0x0001, 
+    0x0007, 0x5055, 0x5348, 0x2044, 0xffff,
+    0x0005, 0x0001, 
+    0x0007, 0x5354, 0x4152, 0x54ff, 
+    0x0005, 0x0001, 
+    0x0007, 0x5345, 0x4c45, 0x4354, 0xffff, 
+    0x0003, 0x7327, 
+    0x0007, 0x3132, 0x3334, 0x3536, 0x3738, 0xffff, 
+    0x0003, 0x7288, 
+    0x0007, 0x4449, 0x50ff, 
+    0x0003, 0x728b,
+    0x0007, 0x5445, 0x5354, 0xffff, 
+    0x0005, 0x0001, 
+    0x0007, 0x434f, 0x494e, 0x31ff, 
+    0x0005, 0x0001, 
+    0x0007, 0x434f, 0x494e, 0x32ff, 
+    0x0005, 0x0001, 
+    0x0007, 0x5345, 0x5256, 0x4943, 0x45ff, 
+    0x0003, 0x7293, 
+    0x0007, 0x454c, 0x204c, 0x4544, 0x3120, 0x4c45, 0x4432, 0xffff,
+    0x0000
+};
+
+static const uint16_t io_values [] = {
+    // EL
+    0x0101, 0x0002,
+    0x2002,
+    0x0003, 0x72b4,
+    0x0004, 0x0010, 0xfc00,
+
+    // LED 1
+    0x0003, 0x7314,
+    0x0006,
+
+    // LED 2
+    0x0003, 0x73b4,
+    0x0006,
+
+    // P1 / P2
+    0x0101, 0x000a,
+    0x0102,
+    0x0003, 0x71a9,
+    0x0006,
+    0x0003, 0x7209,
+    0x0006, 
+
+    // Dip switches
+    0x0101, 0x0008,
+    0x2002,
+    0x0003, 0x7328,
+    0x0006,
+
+    // Test, coin and service
+    0x0101, 0x0004,
+    0x0102,
+    0x0003, 0x73ab,
+    0x0006,
+
+    0x0000
+};
+
+static const uint16_t sound [] = {
+    0x0003, 0x71c6, 
+    0x0108, 0x534f, 0x554e, 0x4420, 0x5445, 0x5354, 0xffff,
+    0x000a, 0x0010, 0xfc00,
+    0x0000, 0x0000
+};
+
+static const char * const sound_menu[] = {
+    "NO SOUND",
+    "RIGHT",
+    "LEFT",
+    "CENTER"
+};
+
+static const uint16_t memory_card [] = {
+    0x0003, 0x7186,
+    0x0108, 0x4d45, 0x4d4f, 0x5259, 0x2043, 0x4152, 0x4420, 0x5445, 0x5354, 0xffff,
+    0x0003, 0x714c,
+    0x1108, 0x1149, 0x4e53, 0x4552, 0x5420, 0x4d45, 0x4d4f, 0x5259, 0x2043, 0x4152, 0x4410, 0xffff, 
+    0x0000
+};
+
+static const uint16_t memory_card_ok [] = {
+    0x0003, 0x70ec, 
+    0x0108, 0x4d45, 0x4d4f, 0x5259, 0x2043, 0x4152, 0x4420, 0x5445, 0x5354, 0x204f, 0x4bff,
+    0x0000
+};
+
+static const uint16_t service_memory_card_error [] = {
+    0x0003, 0x70ec,
+    0x0108, 0x4d45, 0x4d4f, 0x5259, 0x2043, 0x4152, 0x4420, 0x5445, 0x5354, 0x2045, 0x5252, 0x4f52, 0xffff, 
+    0x0000
+};
+
+static const uint16_t backup_clear [] = {
+    // 'backup clear'
+    0x0003, 0x71c6,
+    0x0108, 0x4241, 0x434b, 0x5550, 0x2043, 0x4c45, 0x4152, 0xffff, 
+    // 'ok'
+    0x0003, 0x710c,
+    0x1108, 0x4f4b, 0xffff,
+    // ' = press a,b,c buttons'
+    0x0003, 0x714c,
+    0x0108, 0x203d, 0x2050, 0x5245, 0x5353, 0x2041, 0x2c42, 0x2c43, 0x2042, 0x5554, 0x544f, 0x4e53, 0xffff, 
+    // '     at the same time'
+    0x0003, 0x710e,
+    0x0108, 0x2020, 0x2020, 0x2041, 0x5420, 0x5448, 0x4520, 0x5341, 0x4d45, 0x2054, 0x494d, 0x45ff, 
+    0x0000
+};
+
+static const uint16_t service_setup_calendar [] = {
+    // 'setting up the calendar'
+    0x0003, 0x7106, 
+    0x0108, 0x5345, 0x5454, 0x494e, 0x4720, 0x5550, 0x2054, 0x4845, 0x2043, 0x414c, 0x454e, 0x4441, 0x52ff, 
+
+    // 'current time'
+    0x0003, 0x7169,
+    0x0108, 0x4355, 0x5252, 0x454e, 0x5420, 0x5449, 0x4d45, 0xffff, 
+    
+    // Add /
+    0x0003, 0x71ac,
+    0x0108, 0x2f20, 0x202f, 0xffff,
+
+    // Add :
+    0x0003, 0x72cc,
+    0x0108, 0x3a20, 0x203a, 0xffff, 
+
+    // 'new time'
+    0x0003, 0x7170, 
+    0x0108, 0x4e45, 0x5720, 0x5449, 0x4d45, 0xffff,
+
+    // Add / and : 
+    0x0003, 0x71b3,
+    0x0108, 0x2f20, 0x202f, 0x2020, 0x2020, 0x203a, 0x2020, 0x3aff,
+
+    // 'joystick = select'
+    0x0003, 0x7176,
+    0x0108, 0x4a4f, 0x5953, 0x5449, 0x434b, 0x203d, 0x2053, 0x454c, 0x4543, 0x54ff,
+
+    // 'd button = set'
+    0x0003, 0x7179,
+    0x0108, 0x4420, 0x4255, 0x5454, 0x4f4e, 0x203d, 0x2053, 0x4156, 0x45ff, 
+
+    0x0000
+};
+
+#endif // _SERVICE_HARDWARE_H
