+++ src/backup_ram.c
@@ -40,13 +40,15 @@ void reset_backup_ram() {
   
     // Reset saved slot data
     for (uint8_t i = 0; i < 8; i++) {
-        *BRAM_SLOT_NGH_ID(i) = 0x0000FFFF;
+        BRAM_NGH_BLOCK[i].ngh = 0x0000;
+        BRAM_NGH_BLOCK[i].block = 0xFFFF;
     }
 
     lock_backup_ram();
 }
 
 // Writes additional values into backup-ram. Not part of the reset, but values that are added later on.
+// This function locks backup ram
 void setup_backup_ram() {
     unlock_backup_ram();
 
@@ -58,6 +60,7 @@ void setup_backup_ram() {
     // TODO: Maybe save credit if restarted?
     volatile uint8_t *address = (volatile uint8_t *)BRAM_COIN_COUNTER_INTERNAL;
     for (uint8_t i = 0; i < 22; i++) {
+        *WATCHDOG = 0;
         *address = 0x00;
         address++;
     }
@@ -76,32 +79,141 @@ void setup_backup_ram() {
         }
     }
 
+    // Reset temporary NGH-memory
+    for (uint8_t slot = 0; slot < MAX_SLOTS; slot++) {
+        *WATCHDOG = 0;
+        BIOS_NGH_BLOCK[slot] = (SlotEntry){
+            .ngh = 0x0000,
+            .block = 0xFFFF,
+        };
+    }
+
     // Find first working slot
     *BRAM_FIRST_PLAYABLE_SLOT = 0xFF;
     *REG_SWPROM = 0;
+
     for (uint8_t slot = 0; slot < *BRAM_SLOT_COUNT; slot++) {
+        *WATCHDOG = 0;
         *REG_SLOT = slot;
-        if (*ROM_NGH_NUMER != 0xFFFF) {
+        if (*ROM_NGH_NUMBER != 0xFFFF) {
             if (0xFF == *BRAM_FIRST_PLAYABLE_SLOT) {
                 *BRAM_FIRST_PLAYABLE_SLOT = slot;
             }
+
             // TODO: Verify security string 
-            /*
-            uint16_t block = find_game_data_block_by_name();
-            if (block == 0xFF) {
-                block = find_next_available_data_block();
-            }
-            // TODO: Show a warning if all blocks are taken or maybe just overwrite one?
-            *BRAM_SLOT_NGH_ID(slot) = (*ROM_NGH_NUMER << 16) | block;
-            */
+
+            BIOS_NGH_BLOCK[slot] = (SlotEntry){
+                .ngh = *ROM_NGH_NUMBER,
+                .block = 0xFFFF,
+            };
         }
     }
-    *REG_SWPBIOS = 0;    
+    *BRAM_SLOT_SELECTED = *BRAM_FIRST_PLAYABLE_SLOT;
+    *REG_SLOT = *BRAM_FIRST_PLAYABLE_SLOT;
+   
+    // Check for duplicate games
+    for (uint8_t i = 0; i + 1 < *BRAM_SLOT_COUNT; i++) {
+        *WATCHDOG = 0;
+        uint16_t id = BIOS_NGH_BLOCK[i].ngh;
+        if (id == 0) {
+            continue;
+        }
+        for (uint8_t j = i + 1; j < MAX_SLOTS; j++) {
+            if (BIOS_NGH_BLOCK[j].ngh == id) {
+                BIOS_NGH_BLOCK[j].ngh = 0;
+            }
+        }
+    }
+
+    // Save 'orphan' games into a temporary slot at D00164
+    volatile OrphanBookEntry* orphans  = ORPHAN_BOOKS;
+    for (uint8_t i = 0; i < MAX_SLOTS; i++) {
+        *WATCHDOG = 0;
+        orphans[i].ngh = 0x0000;
+        orphans[i].block = 0xFFFF;
+        orphans[i].created = 0;
+    }
+
+    uint8_t outIndex = 0;
+ 
+    // Loop all BRAM slots and check if those games exist in active BIOS slots
+    for (uint8_t i = 0; i < MAX_SLOTS; i++) {
+        *WATCHDOG = 0;
+        uint16_t ngh = BRAM_NGH_BLOCK[i].ngh;
+        if (ngh == 0x0000) {
+            continue;
+        }
+
+        uint8_t found = 0;
+        for (uint8_t s = 0; s < *BRAM_SLOT_COUNT; s++) {
+            if (BIOS_NGH_BLOCK[s].ngh == ngh) {
+                found = 1;
+                break;
+            }
+        }
+
+        if (found) {
+            continue;
+        }
+
+        if (outIndex >= MAX_SLOTS) {
+            break;
+        }
+
+        orphans[outIndex].ngh = BRAM_NGH_BLOCK[i].ngh;
+        orphans[outIndex].block = BRAM_NGH_BLOCK[i].block;
+        orphans[outIndex].created = *BRAM_SLOT_BOOKKEEP_DATE(i);
+        outIndex++;
+    }
+
+    // Find already assigned blocks for current games and save it to BIOS slots
+    for (uint8_t s = 0; s < *BRAM_SLOT_COUNT; s++) {
+        *WATCHDOG = 0;
+        uint16_t ngh = BIOS_NGH_BLOCK[s].ngh;
+        if (ngh == 0x0000) {
+            continue;
+        }
+
+        uint8_t i;
+        for (i = 0; i < MAX_SLOTS; i++) {
+            if (BRAM_NGH_BLOCK[i].ngh == ngh) {
+                break;
+            }
+        }
+
+        if (i == MAX_SLOTS) {
+            continue;
+        }
+
+        BIOS_NGH_BLOCK[s].block = BRAM_NGH_BLOCK[i].block;
+        *BIOS_SLOT_BOOKKEEP_DATE(s) = *BRAM_SLOT_BOOKKEEP_DATE(i);
+    }
+
+    // Insert BIOS slots data to BRAM
+    for (uint8_t i = 0; i < *BRAM_SLOT_COUNT; i++) {
+        *WATCHDOG = 0;
+        BRAM_NGH_BLOCK[i].ngh = BIOS_NGH_BLOCK[i].ngh;
+        BRAM_NGH_BLOCK[i].block = BIOS_NGH_BLOCK[i].block;
+
+        *BRAM_SLOT_BOOKKEEP_DATE(i) = *BIOS_SLOT_BOOKKEEP_DATE(i);
+    }
+
+    // Append "orphan" games, first come, first served
+    for (uint8_t i = *BRAM_SLOT_COUNT; i < MAX_SLOTS; i++) {
+        *WATCHDOG = 0;
+        BRAM_NGH_BLOCK[i].ngh = orphans->ngh;
+        BRAM_NGH_BLOCK[i].block = orphans->block;
+        *BRAM_SLOT_BOOKKEEP_DATE(i) = orphans->created;
+        orphans++;
+    }
+
+    *REG_SWPBIOS = 0;
     
     lock_backup_ram();
 }
 
 void lock_backup_ram() {
+    // TODO: Investigate what this is actually for.
     if (*BIOS_BRAM_USED) {
         *BIOS_BRAM_USED = 0;
     }
@@ -113,117 +225,89 @@ void unlock_backup_ram() {
     *BACKUP_RAM_UNLOCK = 0;
 }
 
-void init_game_data() {
-    load_game_data();
-
-    *REG_SWPROM = 0;
-    *BIOS_SWPMODE = 0;
-    *BIOS_USER_REQUEST = 0;
-    *BIOS_SYSRET_STATUS = 0;
-    // With BIOS_USER_REQUEST = 0 this initialize the default highscore etc
-    // Will not return here but call the system_return()
-    SUBR_CART_USER();
-}
-
 // Loads game data from workram. Tries to reserve a block if none is assigned yet.
+// This function locks backup ram
 void load_game_data() {
     unlock_backup_ram();
-
-    uint8_t block = find_game_data_block();
-
     *REG_SWPROM = 0;
-    *BIOS_SWPMODE = 0;
 
-    // If we did not find an already reserved block, take the next available and fill it with data
-    if (0xFF == block) {
-        for (uint8_t i = 0; i < 8; i++) {
-            if ((uint16_t) *BRAM_SLOT_NGH_ID(i) != 0x0000) {
-                block = i;
-                break;
-            }
-        }
+    uint16_t rom_ngh = *ROM_NGH_NUMBER;
+    uint16_t block = find_game_data_block(rom_ngh);
 
-        // TODO: Make sure to add an error if there is no more available block
-        if (0xFF == block) {
-            return;
-        }
-
-        // Insert the NGH-number and selected slot
-        *BRAM_SLOT_NGH_ID(block) = (*ROM_NGH_NUMER << 16) | (uint16_t)block;
-
-        // Insert game name and dip settings
-        uint8_t region = *SROM_COUNTRY_CODE;
-        uint32_t dips_addr = ROM_SOFTDIP_TABLE[region];
-        volatile uint8_t *dips = (volatile uint8_t *)(uint32_t)dips_addr;
-
-        // Copy the game name to backup ram block
-        for (uint8_t i = 0; i < 16; i++) {
-            BRAM_GAME_NAME(block)[i] = *dips;
-            dips++;
-        }
-
-        /*
-            Metal Slug default dipsw
-
-            12 38 04 12 12 02 01 04 03 00
-            continue        12      off/on*
-            difficulty      38      1,2,3,4*,5,6,7,8
-            play time       04      60*,70,80,90
-            demo sound      12      off,on*
-            play manual     12      off,on*
-            blood           02      off*,on
-
-        */
-
-        // Copy timed/counter options
-        for (uint8_t i = 0; i < 6; i++) {
-            BRAM_GAME_DIP_SETTINGS(block)[i] = *dips;
-            dips++;
-        }
-        // Get default values for all 10 potential soft dips.
-        for (uint8_t i = 6; i < 16; i++) {
-            BRAM_GAME_DIP_SETTINGS(block)[i] = *dips >> 4;
-            dips++;
-        }
-    }
-
-    // Copy backup-data into workram (where the game expect to find it)
-    volatile uint8_t *bram_game_data = BRAM_GAME_BLOCK(block);
-    volatile uint8_t *rom_game_data = (volatile uint8_t *) *ROM_BACKUP_RAM_PTR;
-
-    volatile uint16_t size = *ROM_GAME_SAVE_SIZE;
-    if (size > 0x1000) {
-        size = (uint16_t)BRAM_GAME_BLOCK_SIZE;
-    }
-
-    for (uint16_t i = 0; i < size; i++) {
-        rom_game_data[i] = bram_game_data[i];
+    // Check if game slot is unassigned
+    if (0xFFFF == block) {
+        goto _load_game_data_end;
     }
 
     // Load dip settings
-    for (uint8_t i = 0; i < 16; i++) {
-        BIOS_GAME_DIP[i] = BRAM_GAME_DIP_SETTINGS(block)[i];
+    memcpy((void*)BIOS_GAME_DIP_PTR, (void*)BRAM_GAME_DIP_SETTINGS(block), 16);
+
+    // Load game block
+    volatile uint16_t size = *ROM_GAME_SAVE_SIZE;
+    if (size > BRAM_GAME_BLOCK_SIZE) {
+        size = BRAM_GAME_BLOCK_SIZE;
     }
+    memcpy((void*)*ROM_BACKUP_RAM_PTR, (void*)BRAM_GAME_BLOCK(block), size);
 
+_load_game_data_end:
     *REG_SWPBIOS = 0;
-    *BIOS_SWPMODE = 0xFF;
-
     lock_backup_ram();
 }
 
 // Saves game data from workram
-// Before calling this function, make sure you set REG_SWPROM
 void save_game_data() {
     unlock_backup_ram();
+    // TODO: Sort our a better way to not fight with bram-lock and vblank routine
+    __asm__ volatile ("move #0x2700, %sr");     // Disable interrupt
+    *REG_SWPROM = 0;
+    *BIOS_SWPMODE = 0x00;
 
-    uint8_t block = find_game_data_block();
+    uint16_t rom_ngh = *ROM_NGH_NUMBER;
+    uint16_t block = find_game_data_block(rom_ngh);
+    uint8_t new_block = 0;
 
-    if (0xFF == block) {
-        return;
+    // Check if game slot is unassigned
+    if (0xFFFF == block) {
+        block = find_next_available_data_block();
+        
+        for (uint8_t bram_slot = 0; bram_slot < MAX_SLOTS; bram_slot++) {
+            if (rom_ngh != BRAM_NGH_BLOCK[bram_slot].ngh) {
+                continue;
+            }
+            BRAM_NGH_BLOCK[bram_slot].block = block;
+            reset_game_data(block);
+            *BRAM_SLOT_BOOKKEEP_DATE(bram_slot) = *(volatile uint32_t *)BRAM_DATE_PTR;
+            new_block = 1;
+            break;
+        }
     }
 
-    *REG_SWPROM = 0;
-    *BIOS_SWPMODE = 0;
+    // Make sure not to overwrite existing block data
+    if (0 == *BIOS_SYSRET_STATUS && 0 == new_block) {
+        goto _save_game_data_end;
+    }
+
+    // Copy metadata
+    uint8_t region = *SROM_COUNTRY_CODE;
+    volatile uint32_t dips_addr = ROM_SOFTDIP_TABLE[region];
+    volatile uint8_t *dips = (volatile uint8_t *)dips_addr;
+
+    // Copy the game name to backup ram block
+    for (uint8_t i = 0; i < 16; i++) {
+        BRAM_GAME_NAME(block)[i] = *dips;
+        dips++;
+    }
+
+    // Copy timed/counter options
+    for (uint8_t i = 0; i < 6; i++) {
+        BRAM_GAME_DIP_SETTINGS(block)[i] = *dips;
+        dips++;
+    }
+    // Copy default values for all 10 potential soft dips.
+    for (uint8_t i = 6; i < 16; i++) {
+        BRAM_GAME_DIP_SETTINGS(block)[i] = *dips >> 4;
+        dips++;
+    }
 
     // Copy game data into backup
     volatile uint8_t *bram_game_data = BRAM_GAME_BLOCK(block);
@@ -234,29 +318,35 @@ void save_game_data() {
         size = (uint16_t)BRAM_GAME_BLOCK_SIZE;
     }
 
-    for (uint16_t i = 0; i < size; i++) {
-        bram_game_data[i] = rom_game_data[i];
-    }
+    memcpy((void*)BRAM_GAME_BLOCK(block), (void*)*ROM_BACKUP_RAM_PTR, size);
 
+_save_game_data_end:
     *REG_SWPBIOS = 0;
     *BIOS_SWPMODE = 0xFF;
-
+    __asm__ volatile ("move #0x2000, %sr");     // Enable interrupt
     lock_backup_ram();
 }
 
-uint8_t find_game_data_block() {
-    *REG_SWPROM = 0;
-    *BIOS_SWPMODE = 0x00;    
-    uint8_t block = 0xFF;
-    for (uint8_t i = 0; i < 8; i++) {
-        uint16_t bram_ngh_id = (uint16_t) ((*BRAM_SLOT_NGH_ID(i) & 0xFFFF0000) >> 16);
-        if (bram_ngh_id == *ROM_NGH_NUMER) {
-            block = i;
+void reset_game_data(uint16_t block) {
+    // Reset book date
+    memset((void *)BRAM_SLOT_BOOKKEEP_DATE(block), 0, 4);
+    // Reset soft dips
+    memset((void *)BRAM_GAME_DIP_SETTINGS(block), 0, 16);
+    // Reset game name
+    memset((void *)BRAM_GAME_NAME(block), 0, 16);
+    // Reset game data
+    memset((void *)BRAM_GAME_BLOCK(block), 0, BRAM_GAME_BLOCK_SIZE);
+}
+
+uint16_t find_game_data_block(uint16_t rom_ngh) {
+    uint16_t block = 0xFFFF;
+    for (uint16_t i = 0; i < BRAM_MAX_SLOTS; i++) {
+
+        if (BRAM_NGH_BLOCK[i].ngh == rom_ngh) {
+            block = BRAM_NGH_BLOCK[i].block;
             break;
         }
     }
-    *REG_SWPBIOS = 0;
-    *BIOS_SWPMODE = 0xFF;    
     return block;
 }
 
@@ -282,19 +372,24 @@ uint16_t find_game_data_block_by_name() {
     return block;
 }
 
+// Finds the next available data block based on current used BRAM_NGH_BLOCK-data.
 uint16_t find_next_available_data_block() {
-    uint16_t block = 0xFF;
-    for (uint8_t i = 0; i < 8; i++) {
-        uint8_t available = 1;
-        for (uint8_t c = 0; c < 0x10; c++) {
-            if (BRAM_GAME_NAME(i)[c] != 0x00) {
-                available = 0;
+    for (uint16_t candidate = 0; candidate < MAX_SLOTS; candidate++) {
+        *WATCHDOG = 0;
+        uint8_t used = 0;
+        for (uint8_t i = 0; i < MAX_SLOTS; i++) {
+            uint16_t block = BRAM_NGH_BLOCK[i].block;
+            if (block == 0xFFFF) {
+                continue;
+            }
+            if ((block & 0x00FF) == (candidate & 0x00FF)) {
+                used = 1;
+                break;
             }
         }
-        if (available == 1) {
-            block = i;
-            break;
+        if (!used) {
+            return candidate;
         }
     }
-    return block;
+    return 0xFFFF;
 }
