+++ src/bios_calls.c
@@ -11,6 +11,9 @@
 #include "input.h"
 #include "interrupt.h"
 #include "mess_out.h"
+#include "service.h"
+#include "service_dips.h"
+#include "service_hardware.h"
 #include "utils.h"
 
 #include "bios_calls.h"
@@ -33,20 +36,49 @@ void system_return(void) {
 
     controller_setup();
 
-    if (0 == *BIOS_SYSRET_STATUS) {
-        *REG_SWPROM = 0;
-        save_game_data();
-        *REG_SWPBIOS = 0;
-
-        *BIOS_SYSRET_STATUS = 3;
-        if (0 == *ROM_EYECATCH_FLAG) {          // Eyecatcher handled by BIOS
-            play_bios_eyecatcher();
-        } else if (1 == *ROM_EYECATCH_FLAG) {   // Eyecatcher handled by Game
-            *BIOS_USER_REQUEST = 0x01;
+    if (0 == *BIOS_SYSRET_STATUS) { // Init
+        if (8 != *BRAM_SLOT_CURSOR) {
+            // save_dips()
+            save_game_data();
+            change_slot_incremental();
+            unlock_backup_ram();
+            if (*BRAM_SLOT_SELECTED <= *BRAM_SLOT_CURSOR) {
+                *BRAM_SLOT_CURSOR = 8;
+            } else {
+                *BRAM_SLOT_CURSOR = *BRAM_SLOT_SELECTED;
+            }
+            lock_backup_ram();
+            // 0 = Initialize the default highscore etc
+            *BIOS_USER_REQUEST = 0;           
         }
-    } else if (3 == *BIOS_SYSRET_STATUS) {
+        if (8 == *BRAM_SLOT_CURSOR) {
+            // Check if service-button is pressed
+            uint8_t service_button = ((~(*REG_STATUS_A)) & 0x4) >> 2;
+            uint8_t dipsw_settings = (~(*REG_DIPSW)) & 0x1;
+            if (service_button || dipsw_settings) {
+                show_bios_menu();
+            }
+
+            // If no game was found (could not read NGH-number), show test-menu
+            if (0xFF == *BRAM_FIRST_PLAYABLE_SLOT) {
+                show_bios_hardware_test();
+            }
+
+            // Normal initialization
+            *REG_SLOT = *BRAM_FIRST_PLAYABLE_SLOT;
+            *BRAM_SLOT_SELECTED = *BRAM_FIRST_PLAYABLE_SLOT;
+            *BIOS_SYSRET_STATUS = 3;
+            load_game_data();
+
+            if (0 == *ROM_EYECATCH_FLAG) {          // Eyecatcher handled by BIOS
+                play_bios_eyecatcher();
+            } else if (1 == *ROM_EYECATCH_FLAG) {   // Eyecatcher handled by Game
+                *BIOS_USER_REQUEST = 0x01;
+            }
+        }
+    } else if (3 == *BIOS_SYSRET_STATUS) {  // Game over
         save_game_data();
-        load_game_data();
+        //load_game_data();
         if (*BRAM_CREDIT_P1 != 0 || *BRAM_CREDIT_P2 != 0) {
             *BIOS_USER_REQUEST = 0x03;
 
@@ -58,6 +90,25 @@ void system_return(void) {
         } else {
             *BIOS_USER_REQUEST = 0x02;
         }
+    } else if (4 == *BIOS_SYSRET_STATUS || 5 == *BIOS_SYSRET_STATUS) {  // Next/previous slot
+        save_game_data();
+        if (4 == *BIOS_SYSRET_STATUS) {
+            change_slot_incremental();
+        } else {
+            change_slot_decremental();
+        }
+
+        load_game_data();
+        *BIOS_USER_MODE = 1;
+        // Ignore eyecatcher when slot is changed
+        *BIOS_USER_REQUEST = 2;
+
+        // Ignore Demo if credit is present
+        if (*BRAM_CREDIT_P1 != 0 || *BRAM_CREDIT_P2 != 0) {
+            *BIOS_USER_REQUEST = 3;
+        }
+
+        *BIOS_NEXT_GAME_ROTATE = 0;
     }
 
     start_game();
@@ -87,6 +138,9 @@ void system_io(void) {
                 *BIOS_FRAME_SKIP = 1;           // Delay fix for something?
                 // *BIOS_SYSRET_STATUS = 0x03;     // Game change this value to 2 after compulsion ends
             }
+
+            // Check if 'select' is pressed to change the game on a multislot
+            check_change_game();
         }
     }
     
@@ -104,11 +158,11 @@ void system_io(void) {
 
         lock_backup_ram();
     }
-
+/*
     if (menu_hotkey_pressed()) {
         menu_toggle();
     }
-    
+*/  
     return;
 }
